# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CKB Agent Wallet Library API — programmable wallet for AI agents
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type AgentWalletNetwork = typeof AgentWalletNetwork[keyof typeof AgentWalletNetwork];


export const AgentWalletNetwork = {
  mainnet: 'mainnet',
  testnet: 'testnet',
} as const;

export type AgentWalletStatus = typeof AgentWalletStatus[keyof typeof AgentWalletStatus];


export const AgentWalletStatus = {
  active: 'active',
  killed: 'killed',
} as const;

export interface AgentWallet {
  id: number;
  name: string;
  /** CKB address (secp256k1-based) */
  address: string;
  /** Hex-encoded compressed public key */
  publicKey: string;
  network: AgentWalletNetwork;
  status: AgentWalletStatus;
  isKilled: boolean;
  createdAt: string;
  updatedAt: string;
}

export type WalletInputNetwork = typeof WalletInputNetwork[keyof typeof WalletInputNetwork];


export const WalletInputNetwork = {
  mainnet: 'mainnet',
  testnet: 'testnet',
} as const;

export interface WalletInput {
  /** @minLength 1 */
  name: string;
  network: WalletInputNetwork;
  /** Passphrase for encrypting the private key. If omitted, uses server default. */
  passphrase?: string;
}

export interface WalletBalance {
  walletId: number;
  address: string;
  /** Total capacity in shannons (1 CKB = 100,000,000 shannons) */
  totalCapacity: string;
  /** Spendable capacity in shannons */
  freeCapacity: string;
  cellCount: number;
}

export type ScriptHashType = typeof ScriptHashType[keyof typeof ScriptHashType];


export const ScriptHashType = {
  data: 'data',
  type: 'type',
  data1: 'data1',
  data2: 'data2',
} as const;

export interface Script {
  codeHash: string;
  hashType: ScriptHashType;
  args: string;
}

export interface Cell {
  txHash: string;
  index: string;
  /** Capacity in shannons */
  capacity: string;
  lockScript: Script;
  typeScript?: Script;
  /** Hex-encoded cell data */
  data?: string;
}

export interface TransferInput {
  /** Recipient CKB address */
  toAddress: string;
  /** Amount in shannons */
  amount: string;
  /** Optional memo attached as cell data */
  memo?: string;
}

/**
 * Raw CKB transaction JSON to sign
 */
export type SignTxInputRawTx = { [key: string]: unknown };

export interface SignTxInput {
  /** Raw CKB transaction JSON to sign */
  rawTx: SignTxInputRawTx;
}

export type TxResultStatus = typeof TxResultStatus[keyof typeof TxResultStatus];


export const TxResultStatus = {
  pending: 'pending',
  committed: 'committed',
  rejected: 'rejected',
} as const;

export interface TxResult {
  txHash: string;
  status: TxResultStatus;
  /** @nullable */
  message?: string | null;
}

export type TxRecordDetails = { [key: string]: unknown };

export interface TxRecord {
  id: number;
  walletId: number;
  action: string;
  /** @nullable */
  txHash?: string | null;
  /** @nullable */
  amountShannons?: string | null;
  status: string;
  details?: TxRecordDetails;
  createdAt: string;
}

export interface SafetyRules {
  id: number;
  walletId: number;
  /**
     * Max CKB per single transfer (shannons)
     * @nullable
     */
  maxTransferAmountShannons?: string | null;
  /**
     * Max daily spending in shannons
     * @nullable
     */
  dailySpendingLimitShannons?: string | null;
  /** Amount spent today in shannons */
  dailySpentShannons?: string;
  /** Allowed actions: transfer, sign, fiber_open, fiber_pay, dob_mint, otx_compose */
  allowedActions: string[];
  /** If non-empty, only these addresses can receive funds */
  addressWhitelist?: string[];
  isActive: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface SafetyRulesUpdate {
  /** @nullable */
  maxTransferAmountShannons?: string | null;
  /** @nullable */
  dailySpendingLimitShannons?: string | null;
  allowedActions?: string[];
  addressWhitelist?: string[];
  isActive?: boolean;
}

export type FiberChannelStatus = typeof FiberChannelStatus[keyof typeof FiberChannelStatus];


export const FiberChannelStatus = {
  pending: 'pending',
  open: 'open',
  closing: 'closing',
  closed: 'closed',
} as const;

export interface FiberChannel {
  id: number;
  walletId: number;
  /**
     * On-chain channel ID (hex)
     * @nullable
     */
  channelId?: string | null;
  peerAddress: string;
  localCapacityShannons: string;
  /** @nullable */
  remoteCapacityShannons?: string | null;
  status: FiberChannelStatus;
  /** @nullable */
  txHash?: string | null;
  createdAt: string;
}

export interface FiberChannelInput {
  /** Peer node address (multiaddr format) */
  peerAddress: string;
  /** Local commitment capacity in shannons */
  localCapacityShannons: string;
}

export interface FiberPaymentInput {
  /** Fiber payment invoice (fiber:...) */
  invoice: string;
  amountShannons: string;
  /**
     * Optional specific channel to use
     * @nullable
     */
  channelId?: string | null;
}

export type DobAssetStatus = typeof DobAssetStatus[keyof typeof DobAssetStatus];


export const DobAssetStatus = {
  pending: 'pending',
  minted: 'minted',
  burned: 'burned',
} as const;

export interface DobAsset {
  id: number;
  walletId: number;
  /**
     * On-chain token ID
     * @nullable
     */
  tokenId?: string | null;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  contentType?: string | null;
  /**
     * Base64-encoded content or URI
     * @nullable
     */
  content?: string | null;
  /** @nullable */
  clusterName?: string | null;
  /** @nullable */
  txHash?: string | null;
  status: DobAssetStatus;
  createdAt: string;
}

export interface DobMintInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  /** MIME type, e.g. text/plain, image/png */
  contentType?: string;
  /** Base64-encoded content or URI */
  content?: string;
  /** Optional cluster/collection name */
  clusterName?: string;
}

export type OtxIntentIntentData = { [key: string]: unknown };

export type OtxIntentStatus = typeof OtxIntentStatus[keyof typeof OtxIntentStatus];


export const OtxIntentStatus = {
  pending: 'pending',
  finalized: 'finalized',
  failed: 'failed',
} as const;

export interface OtxIntent {
  id: number;
  walletId: number;
  /** e.g. swap, transfer, lock */
  intentType: string;
  intentData?: OtxIntentIntentData;
  status: OtxIntentStatus;
  /** @nullable */
  txHash?: string | null;
  createdAt: string;
}

/**
 * Partial transaction data for composing the OTX
 */
export type OtxIntentInputIntentData = { [key: string]: unknown };

export interface OtxIntentInput {
  intentType: string;
  /** Partial transaction data for composing the OTX */
  intentData: OtxIntentInputIntentData;
}

export type AuditLogEntryStatus = typeof AuditLogEntryStatus[keyof typeof AuditLogEntryStatus];


export const AuditLogEntryStatus = {
  success: 'success',
  blocked: 'blocked',
  failed: 'failed',
} as const;

export type AuditLogEntryDetails = { [key: string]: unknown };

export interface AuditLogEntry {
  id: number;
  walletId: number;
  action: string;
  /** @nullable */
  txHash?: string | null;
  /** @nullable */
  amountShannons?: string | null;
  status: AuditLogEntryStatus;
  /** @nullable */
  blockedReason?: string | null;
  details?: AuditLogEntryDetails;
  createdAt: string;
}

export interface DashboardStats {
  totalWallets: number;
  activeWallets: number;
  killedWallets: number;
  totalTransactions: number;
  blockedTransactions: number;
  openChannels: number;
  mintedDobs: number;
}