# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CKB Agent Wallet Library API — programmable wallet for AI agents
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /wallets/{id}/cells:
    get:
      operationId: listCells
      tags: [cells]
      summary: List live cells owned by the wallet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of cells
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Cell"
        "404":
          description: Wallet not found

  # ── Transactions ──────────────────────────────────────────────────────────
...
  /wallets/{id}/fiber/channels:
    get:
      operationId: listFiberChannels
      tags: [fiber]
      summary: List Fiber payment channels
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of channels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FiberChannel"
    post:
      operationId: openFiberChannel
      tags: [fiber]
      summary: Open a Fiber payment channel
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FiberChannelInput"
      responses:
        "201":
          description: Channel opened
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FiberChannel"
  /wallets/{id}/fiber/channels/{channelId}:
    delete:
      operationId: closeFiberChannel
      tags: [fiber]
      summary: Close a Fiber payment channel
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: channelId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Channel closed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FiberChannel"
        "404":
          description: Channel not found
  /wallets/{id}/fiber/pay:
    post:
      operationId: fiberPay
      tags: [fiber]
      summary: Make a Fiber network payment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FiberPaymentInput"
      responses:
        "200":
          description: Payment sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TxResult"

  # ── DOBs ──────────────────────────────────────────────────────────────────
  /wallets/{id}/dobs:
    get:
      operationId: listDobs
      tags: [dobs]
      summary: List minted Digital Objects owned by wallet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of DOBs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DobAsset"
    post:
      operationId: mintDob
      tags: [dobs]
      summary: Mint a new Digital Object (DOB/NFT)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DobMintInput"
      responses:
        "201":
          description: DOB minted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DobAsset"

  # ── OTX Intents ───────────────────────────────────────────────────────────
  /wallets/{id}/otx:
    get:
      operationId: listOtxIntents
      tags: [otx]
      summary: List OTX intents
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of OTX intents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OtxIntent"
    post:
      operationId: composeOtxIntent
      tags: [otx]
      summary: Compose an Open Transaction intent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtxIntentInput"
      responses:
        "201":
          description: Intent composed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OtxIntent"
  /wallets/{id}/otx/{intentId}/finalize:
    post:
      operationId: finalizeOtxIntent
      tags: [otx]
      summary: Finalize and broadcast an OTX intent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: intentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Intent finalized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OtxIntent"
        "404":
          description: Intent not found

  # ── Audit Log ─────────────────────────────────────────────────────────────
  /wallets/{id}/audit:
    get:
      operationId: listAuditLog
      tags: [audit]
      summary: Get audit log for wallet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Audit log entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditLogEntry"
        "404":
          description: Wallet not found

  # ── Stats (dashboard) ─────────────────────────────────────────────────────
...
components:
  schemas:
...
    Script:
      type: object
      required: [codeHash, hashType, args]
      properties:
        codeHash:
          type: string
        hashType:
          type: string
          enum: [data, type, data1, data2]
        args:
          type: string
    Cell:
      type: object
      required: [txHash, index, capacity, lockScript]
      properties:
        txHash:
          type: string
        index:
          type: string
        capacity:
          type: string
          description: Capacity in shannons
        lockScript:
          $ref: "#/components/schemas/Script"
        typeScript:
          $ref: "#/components/schemas/Script"
        data:
          type: string
          description: Hex-encoded cell data

    # ── Transactions ────────────────────────────────────────────────────────
...
    TxResult:
      type: object
      required: [txHash, status]
      properties:
        txHash:
          type: string
        status:
          type: string
          enum: [pending, committed, rejected]
        message:
          type: ["string", "null"]
...
    FiberChannel:
      type: object
      required: [id, walletId, peerAddress, localCapacityShannons, status, createdAt]
      properties:
        id:
          type: integer
        walletId:
          type: integer
        channelId:
          type: ["string", "null"]
          description: On-chain channel ID (hex)
        peerAddress:
          type: string
        localCapacityShannons:
          type: string
        remoteCapacityShannons:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, open, closing, closed]
        txHash:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    FiberChannelInput:
      type: object
      required: [peerAddress, localCapacityShannons]
      properties:
        peerAddress:
          type: string
          description: Peer node address (multiaddr format)
        localCapacityShannons:
          type: string
          description: Local commitment capacity in shannons
    FiberPaymentInput:
      type: object
      required: [invoice, amountShannons]
      properties:
        invoice:
          type: string
          description: Fiber payment invoice (fiber:...)
        amountShannons:
          type: string
        channelId:
          type: ["string", "null"]
          description: Optional specific channel to use

    # ── DOBs ────────────────────────────────────────────────────────────────
    DobAsset:
      type: object
      required: [id, walletId, name, status, createdAt]
      properties:
        id:
          type: integer
        walletId:
          type: integer
        tokenId:
          type: ["string", "null"]
          description: On-chain token ID
        name:
          type: string
        description:
          type: ["string", "null"]
        contentType:
          type: ["string", "null"]
        content:
          type: ["string", "null"]
          description: Base64-encoded content or URI
        clusterName:
          type: ["string", "null"]
        txHash:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, minted, burned]
        createdAt:
          type: string
          format: date-time
    DobMintInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        contentType:
          type: string
          description: "MIME type, e.g. text/plain, image/png"
        content:
          type: string
          description: Base64-encoded content or URI
        clusterName:
          type: string
          description: Optional cluster/collection name

    # ── OTX ─────────────────────────────────────────────────────────────────
    OtxIntent:
      type: object
      required: [id, walletId, intentType, status, createdAt]
      properties:
        id:
          type: integer
        walletId:
          type: integer
        intentType:
          type: string
          description: "e.g. swap, transfer, lock"
        intentData:
          type: object
          additionalProperties: true
        status:
          type: string
          enum: [pending, finalized, failed]
        txHash:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    OtxIntentInput:
      type: object
      required: [intentType, intentData]
      properties:
        intentType:
          type: string
        intentData:
          type: object
          additionalProperties: true
          description: Partial transaction data for composing the OTX

    # ── Audit Log ───────────────────────────────────────────────────────────
    AuditLogEntry:
      type: object
      required: [id, walletId, action, status, createdAt]
      properties:
        id:
          type: integer
        walletId:
          type: integer
        action:
          type: string
        txHash:
          type: ["string", "null"]
        amountShannons:
          type: ["string", "null"]
        status:
          type: string
          enum: [success, blocked, failed]
        blockedReason:
          type: ["string", "null"]
        details:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time

    # ── Dashboard ───────────────────────────────────────────────────────────
...