# lib/api-zod/src/generated/types/otxIntentInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CKB Agent Wallet Library API — programmable wallet for AI agents
 * OpenAPI spec version: 0.1.0
 */
import type { OtxIntentInputIntentData } from './otxIntentInputIntentData';

export interface OtxIntentInput {
  intentType: string;
  /** Partial transaction data for composing the OTX */
  intentData: OtxIntentInputIntentData;
}